# perf: fix misc inefficiencies in bitfield.cc