Noticed when using `transmission-gtk` a large amount of my systems total CPU time is spent in `tr_bitfield::countFlags`.

Optimize the code a bit, a better long term approach for most x86_64 machines would be to use the `popcnt` instruction (accessible via `std::popcount` for std >= c++20 or some library like [libpopcnt](https://github.com/kimwalisch/libpopcnt)). But for now this change get a bit of a speedup that will hopefully help.

All unit tests pass:

```
$> ./libtransmission-test 
[----------] 5 tests from Bitfield
[ RUN      ] Bitfield.count
[       OK ] Bitfield.count (69 ms)
[ RUN      ] Bitfield.ctorFromFlagArray
[       OK ] Bitfield.ctorFromFlagArray (0 ms)
[ RUN      ] Bitfield.setRaw
[       OK ] Bitfield.setRaw (0 ms)
[ RUN      ] Bitfield.bitfields
[       OK ] Bitfield.bitfields (0 ms)
[ RUN      ] Bitfield.hasAllNone
[       OK ] Bitfield.hasAllNone (0 ms)
[----------] 5 tests from Bitfield (69 ms total)
...
[----------] Global test environment tear-down
[==========] 264 tests from 36 test suites ran. (17222 ms total)
[  PASSED  ] 246 tests.
[  SKIPPED ] 18 tests, listed below:
[  SKIPPED ] JSON/JSONTest.testElements/1
[  SKIPPED ] JSON/JSONTest.testElements/2
[  SKIPPED ] JSON/JSONTest.testElements/3
[  SKIPPED ] JSON/JSONTest.testUtf8/1
[  SKIPPED ] JSON/JSONTest.testUtf8/2
[  SKIPPED ] JSON/JSONTest.testUtf8/3
[  SKIPPED ] JSON/JSONTest.test1/1
[  SKIPPED ] JSON/JSONTest.test1/2
[  SKIPPED ] JSON/JSONTest.test1/3
[  SKIPPED ] JSON/JSONTest.test2/1
[  SKIPPED ] JSON/JSONTest.test2/2
[  SKIPPED ] JSON/JSONTest.test2/3
[  SKIPPED ] JSON/JSONTest.test3/1
[  SKIPPED ] JSON/JSONTest.test3/2
[  SKIPPED ] JSON/JSONTest.test3/3
[  SKIPPED ] JSON/JSONTest.unescape/1
[  SKIPPED ] JSON/JSONTest.unescape/2
[  SKIPPED ] JSON/JSONTest.unescape/3
```

(Assuming the skipped tests are fine).


Used the code [here](https://github.com/goldsteinn/transmission/blob/bitfield-bench/tests/libtransmission/bench-test.cc) for benchmarking and got roughly a 10% speedup.

Note for larger range the code can be sped up a fair bit further by actually unrolling the loop but that causes some regressions
for the smaller sizes so didn't go with that approach.

```
Results geometric mean of N = 10 runs
Bench CPU: Tigerlake (11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz)

Geometric Mean of all benchmarks New / Old: 0.905

 Todo, Type, Begin,  End, New Time (ns), Old Time (ns), New Time / Old Time
FIXED, TPUT,     0,    1,         0.078,         0.090,               0.875
FIXED, TPUT,     0,    8,         0.073,         0.076,               0.959
FIXED, TPUT,     0,   15,         0.087,         0.092,               0.940
FIXED, TPUT,     0,   22,         0.093,         0.100,               0.931
FIXED, TPUT,     0,   29,         0.099,         0.107,               0.924
FIXED, TPUT,     0,   36,         0.106,         0.113,               0.937
FIXED, TPUT,     0,   43,         0.114,         0.121,               0.937
FIXED, TPUT,     0,   50,         0.120,         0.129,               0.931
FIXED, TPUT,     0,   57,         0.129,         0.140,               0.923
FIXED, TPUT,     0,   64,         0.130,         0.140,               0.928
FIXED, TPUT,     0,  320,         0.530,         0.570,               0.929
FIXED, TPUT,     0,  576,         0.839,         0.924,               0.908
FIXED, TPUT,     0,  832,         1.111,         1.228,               0.904
FIXED, TPUT,     0, 1088,         1.384,         1.533,               0.902
FIXED, TPUT,     0, 1344,         1.650,         1.827,               0.903
FIXED, TPUT,     0, 1600,         1.924,         2.145,               0.897
FIXED, TPUT,     0, 1856,         2.211,         2.438,               0.907
FIXED, TPUT,     0, 2112,         2.484,         2.748,               0.904
FIXED, TPUT,     0, 2368,         2.744,         3.054,               0.898
FIXED, TPUT,     0, 2624,         3.010,         3.360,               0.896
FIXED, TPUT,     0, 2880,         3.280,         3.667,               0.894
FIXED, TPUT,     0, 3136,         3.550,         3.971,               0.894
FIXED, TPUT,     0, 3392,         3.825,         4.269,               0.896
FIXED, TPUT,     0, 3648,         4.083,         4.575,               0.892
FIXED, TPUT,     0, 3904,         4.358,         4.877,               0.894
FIXED, TPUT,     0, 4160,         4.639,         5.183,               0.895
FIXED, TPUT,     0, 4416,         4.904,         5.484,               0.894
FIXED, TPUT,     0, 4672,         5.172,         5.787,               0.894
FIXED, TPUT,     0, 4928,         5.443,         6.091,               0.894
FIXED, TPUT,     0, 5184,         5.706,         6.409,               0.890
FIXED, TPUT,     0, 5440,         5.987,         6.708,               0.892
FIXED, TPUT,     0, 5696,         6.253,         7.028,               0.890
FIXED, TPUT,     0, 5952,         6.547,         7.318,               0.895
FIXED, TPUT,     0, 6208,         6.800,         7.620,               0.892
FIXED, TPUT,     0, 6464,         7.065,         7.939,               0.890
FIXED, TPUT,     0, 6720,         7.352,         8.223,               0.894
FIXED, TPUT,     0, 6976,         7.651,         8.542,               0.896
FIXED, TPUT,     0, 7232,         7.892,         8.859,               0.891
FIXED, TPUT,     0, 7488,         8.186,         9.157,               0.894
FIXED, TPUT,     0, 7744,         8.459,         9.462,               0.894
FIXED, TPUT,     0, 8000,         8.731,         9.744,               0.896
FIXED, TPUT,     0, 8192,         8.918,         9.982,               0.893
FIXED, TPUT,     1,    2,         0.072,         0.077,               0.928
FIXED, TPUT,     1,    9,         0.085,         0.090,               0.952
FIXED, TPUT,     1,   16,         0.086,         0.090,               0.949
FIXED, TPUT,     1,   23,         0.093,         0.101,               0.922
FIXED, TPUT,     1,   30,         0.099,         0.106,               0.930
FIXED, TPUT,     1,   37,         0.105,         0.113,               0.930
FIXED, TPUT,     1,   44,         0.113,         0.121,               0.935
FIXED, TPUT,     1,   51,         0.119,         0.129,               0.922
FIXED, TPUT,     1,   58,         0.129,         0.138,               0.934
FIXED, TPUT,     1,   65,         0.140,         0.148,               0.945
FIXED, TPUT,     1,  321,         0.555,         0.625,               0.888
FIXED, TPUT,     1,  577,         0.845,         0.932,               0.907
FIXED, TPUT,     1,  833,         1.119,         1.247,               0.898
FIXED, TPUT,     1, 1089,         1.398,         1.544,               0.905
FIXED, TPUT,     1, 1345,         1.662,         1.840,               0.903
FIXED, TPUT,     1, 1601,         1.928,         2.159,               0.893
FIXED, TPUT,     1, 1857,         2.205,         2.457,               0.898
FIXED, TPUT,     1, 2113,         2.493,         2.759,               0.904
FIXED, TPUT,     1, 2369,         2.739,         3.067,               0.893
FIXED, TPUT,     1, 2625,         3.025,         3.376,               0.896
FIXED, TPUT,     1, 2881,         3.295,         3.668,               0.898
FIXED, TPUT,     1, 3137,         3.568,         3.984,               0.896
FIXED, TPUT,     1, 3393,         3.802,         4.277,               0.889
FIXED, TPUT,     1, 3649,         4.086,         4.592,               0.890
FIXED, TPUT,     1, 3905,         4.366,         4.899,               0.891
FIXED, TPUT,     1, 4161,         4.637,         5.201,               0.892
FIXED, TPUT,     1, 4417,         4.927,         5.503,               0.895
FIXED, TPUT,     1, 4673,         5.193,         5.804,               0.895
FIXED, TPUT,     1, 4929,         5.445,         6.107,               0.891
FIXED, TPUT,     1, 5185,         5.712,         6.416,               0.890
FIXED, TPUT,     1, 5441,         6.005,         6.719,               0.894
FIXED, TPUT,     1, 5697,         6.258,         7.018,               0.892
FIXED, TPUT,     1, 5953,         6.533,         7.306,               0.894
FIXED, TPUT,     1, 6209,         6.815,         7.617,               0.895
FIXED, TPUT,     1, 6465,         7.072,         7.941,               0.891
FIXED, TPUT,     1, 6721,         7.346,         8.252,               0.890
FIXED, TPUT,     1, 6977,         7.627,         8.596,               0.887
FIXED, TPUT,     1, 7233,         7.899,         8.926,               0.885
FIXED, TPUT,     1, 7489,         8.178,         9.225,               0.887
FIXED, TPUT,     1, 7745,         8.488,         9.547,               0.889
FIXED, TPUT,     1, 8001,         8.729,         9.829,               0.888
FIXED, TPUT,     1, 8192,         8.921,        10.034,               0.889
 RAND, TPUT,     0,    8,         5.417,         5.402,               1.003
 RAND,  LAT,     0,    8,        11.090,        11.429,               0.970
 RAND, TPUT,     0,   16,         5.244,         5.352,               0.980
 RAND,  LAT,     0,   16,        10.788,        11.280,               0.956
 RAND, TPUT,     0,   32,         4.562,         6.900,               0.661
 RAND,  LAT,     0,   32,        10.103,        12.419,               0.814
 RAND, TPUT,     0,   64,         7.415,         8.396,               0.883
 RAND,  LAT,     0,   64,        12.359,        13.579,               0.910
 RAND, TPUT,     0,  128,         9.811,        10.101,               0.971
 RAND,  LAT,     0,  128,        14.486,        15.122,               0.958
 RAND, TPUT,     0,  256,        12.109,        12.928,               0.937
 RAND,  LAT,     0,  256,        17.058,        17.871,               0.954
 RAND, TPUT,     0,  512,        16.306,        18.160,               0.898
 RAND,  LAT,     0,  512,        21.768,        22.968,               0.948
 RAND, TPUT,     0, 1024,        24.999,        27.917,               0.895
 RAND,  LAT,     0, 1024,        30.527,        32.554,               0.938
 RAND, TPUT,     0, 2048,        42.051,        47.104,               0.893
 RAND,  LAT,     0, 2048,        47.606,        51.835,               0.918
 RAND, TPUT,     0, 4096,        77.179,        86.195,               0.895
 RAND,  LAT,     0, 4096,        82.713,        91.238,               0.907
 RAND, TPUT,     0, 8192,       147.356,       164.277,               0.897
 RAND,  LAT,     0, 8192,       152.995,       169.631,               0.902

```