@ckerr 2022-04-17
I had no idea `libpopcnt` existed. This sounds ideal for our use case in C++17!

Related: https://github.com/transmission/transmission/issues/2830