@goldsteinn 2022-04-17
> I had no idea `libpopcnt` existed. This sounds ideal for our use case in C++17!
> 
> Related: #2830

Yeah might be worth hooking in. Also might be worth expanding the bitfield from using `uint8_t` as its core to `uint32_t` if you do so. Don't think it would cost any extra memory given the current `std::vector` impl as both `uint8_t` and `uint32_t` as factor of `sizeof(std::vector<T>)` (if there is a 'small vec' optimization like `std::string`) and I don't know of any `malloc` impl that doesn't have 4+ byte alignment.

If you hook it in mind pinging back on the PR?