@ckerr 2022-04-18
@goldsteinn yeah I'd be fine with it, sounds good to me.

Also, check out https://sonarcloud.io/project/issues?resolved=false&sinceLeakPeriod=true&types=BUG&id=transmission_transmission&open=AYA5ncZSlbqSPYHOdgf6