@goldsteinn 2022-04-18
> @goldsteinn yeah I'd be fine with it, sounds good to me. Once we're not using a 256-byte lookup table, access should be faster if we widen the vector type to a natural uint.

Working on widening right now. Just a bit of issues to sort out with `raw`/`setRaw`.

> 
> Also, check out https://sonarcloud.io/project/issues?resolved=false&sinceLeakPeriod=true&types=BUG&id=transmission_transmission&open=AYA5ncZSlbqSPYHOdgf6

I'll fix in commit adding libpopcnt. I don't think its a real bug (used `-byte_count` because its fastest way to get the shift we want on most arch).