@reardonia 2022-04-20
typo: "gurantee" -> "guarantee"